les nouvelle changement

+++ backend/src/main/java/net/javaguides/springboot/service/IEntiteService.java
@@ -3,5 +3,5 @@ package net.javaguides.springboot.service;
 import net.javaguides.springboot.model.Entite;
 
 public interface IEntiteService extends IService<Entite, Long>{
-    void deleteById(Long facultyId);
+    void deleteById(Integer facultyId);
 }
